refactor: Extract OnboardingScaffold to their own files
